Consolidate the Jira8 module into main
