--- docs/jira8.rst
@@ -1,4 +1,0 @@
-Jira 8 module
-=============
-
-This is the test module.
